add more caching and handling to the yt module. - webhook queue - template string parsing - caching - general code improvements

+++ backend/Hookio.Contracts/Embeds/Embed.cs
@@ -16,7 +16,7 @@ namespace Hookio.Contracts.Embed
 
         public DateTime? Timestamp { get; set; }
 
-        public decimal? Color { get; set; }
+        public uint? Color { get; set; }
 
         public Footer? Footer { get; set; }
